backend/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
idna==3.4
motor==3.3.1
pydantic==2.3.0
pydantic_core==2.6.3
pymongo==4.5.0
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
